+++ paddle/phi/ops/yaml/inconsistent/static_backward.yaml
@@ -118,7 +118,6 @@
   kernel:
     func: c_softmax_with_cross_entropy_grad
     data_type: loss_grad
-    param: [softmax, label, loss_grad, ignore_index, rank, nranks]
 
 - backward_op : divide_double_grad
   forward : divide_grad (Tensor x, Tensor y, Tensor out, Tensor grad_out, int axis = -1) -> Tensor(grad_x), Tensor(grad_y)
